feature[component]: extracted CartItemComponent to separate file
